Correct Documented Changes Event
Hello, 

Great project by the way. I've been learning about CRDTs and Yjs for a couple of days now and the docs have been really helpful. 

I noticed that documented resulting changes from the transaction in the example in this file might be wrong. In the `Y.Map`, we initially set the value of `food` to `pizza`. So when the transaction runs to update `food` to `pencake` and to add `number` to `31`, shouldn't the `oldValue` for `food` should be `pizza`?

If that's the case, I made the correction in this PR. If it's not the case, it would be great to learn about why and perhaps the docs can be updated to explain. 

Thank you